SYNC-339: Set fields in objects directly if there is no setter.
This enables Orders to be synced in OpenMRS 1.10. Orders have an order_number field, but don't have
a setter for the order number, because order numbers are read-only. Previously, the sync module
would die with a `NullPointerException` when this was the case. Now, we fall back to writing the
field directly if the setter can't be found.

We fall back to directly writing the field instead of making it the default behavior because making
it default behavior causes some tests to fail for a reason I couldn't determine.